requirements.txt
mysql-connector-python==8.0.26
numpy
pandas
datetime
time